+++ .gitignore
@@ -120,3 +120,6 @@ a.md
 
 # Ignore a gunicorn config file
 gunicorn.conf.py
+
+# ignore version file
+src/databricks/labs/mcp/_version.py
