--- servers/unity_catalog/src/unitycatalog_mcp/__main__.py
@@ -1,5 +1,0 @@
-from unitycatalog_mcp import main
-
-
-if __name__ == "__main__":
-    main()
