--- servers/unity_catalog/src/unitycatalog_mcp/version.py
@@ -1,3 +1,0 @@
-import importlib.metadata
-
-VERSION = importlib.metadata.version("unitycatalog-mcp")
