+++ src/databricks/labs/mcp/servers/unity_catalog/__init__.py
@@ -1,7 +1,7 @@
 import sys
 from traceback import format_exc
-from unitycatalog_mcp.server import start
-from unitycatalog_mcp.version import VERSION
+from databricks.labs.mcp.servers.unity_catalog.server import start
+from databricks.labs.mcp._version import __version__ as VERSION
 
 from databricks.sdk.config import with_user_agent_extra
