+++ src/databricks/labs/mcp/servers/unity_catalog/__main__.py
@@ -1,0 +1,5 @@
+from databricks.labs.mcp.servers.unity_catalog import main
+
+
+if __name__ == "__main__":
+    main()
