+++ src/databricks/labs/mcp/servers/unity_catalog/tools/vector_search.py
@@ -7,7 +7,7 @@ from databricks.sdk import WorkspaceClient
 
 from mcp.types import Tool as ToolSpec, TextContent
 
-from unitycatalog_mcp.tools.base_tool import BaseTool
+from databricks.labs.mcp.servers.unity_catalog.tools.base_tool import BaseTool
 
 
 class VectorSearchTool(BaseTool):
