+++ tests/conftest.py
@@ -1,5 +1,5 @@
 import pytest
-from unitycatalog_mcp.cli import get_settings, CliSettings
+from databricks.labs.mcp.servers.unity_catalog.cli import get_settings, CliSettings
 
 
 @pytest.fixture(autouse=True)
